+++ packages/deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/deploy-config-sub-generator
 
+## 0.0.13
+
+### Patch Changes
+
+-   @sap-ux/odata-service-inquirer@2.2.24
+
+## 0.0.12
+
+### Patch Changes
+
+-   1bca2d8: Adapt Deploy Config Inquirer and Deploy Config Sub Generator to support ADP Projects
+-   Updated dependencies [1bca2d8]
+    -   @sap-ux/abap-deploy-config-sub-generator@0.0.54
+
 ## 0.0.11
 
 ### Patch Changes