+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/preview-middleware
 
+## 0.19.1
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.13.18
+
+## 0.19.0
+
+### Minor Changes
+
+-   6e32009: feat: introduce enhanced flp homepage
+    -   controlled via boolean property `flp.enhancedHomePage`, which is false by default
+
 ## 0.18.23
 
 ### Patch Changes