+++ packages/ui-service-sub-generator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui-service-sub-generator
 
+## 0.0.6
+
+### Patch Changes
+
+-   @sap-ux/ui-service-inquirer@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+-   @sap-ux/ui-service-inquirer@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes